refactor: centralise emotion recommendation type
