--- src/types/chat.d.ts
@@ -1,41 +1,0 @@
-
-export interface ChatMessage {
-  id: string;
-  text?: string;
-  content?: string; // Added for compatibility 
-  sender: 'user' | 'assistant' | 'system';
-  role?: 'user' | 'assistant' | 'system'; // For compatibility with different implementations
-  timestamp: string;
-  conversation_id?: string; // For compatibility 
-  conversationId?: string;
-  emotions?: Record<string, number>;
-  feedback?: string;
-  isOptimistic?: boolean;
-  metadata?: Record<string, any>;
-}
-
-export interface ChatConversation {
-  id: string;
-  title: string;
-  updated_at: string;
-  created_at: string;
-  updatedAt?: string; // For compatibility
-  createdAt?: string; // For compatibility
-  last_message?: string;
-  lastMessage?: string; // For compatibility
-  last_message_time?: string;
-  messages?: ChatMessage[];
-  user_id?: string;
-  unread?: number;
-  category?: string;
-  status?: 'active' | 'archived' | 'deleted';
-  metadata?: Record<string, any>;
-}
-
-export interface ChatResponse {
-  message: string;
-  emotion?: string;
-  status: 'success' | 'error';
-  recommendations?: any[];
-  suggestions?: string[];
-}
