+++ src/types/user.ts
@@ -1,8 +1,8 @@
 
 /**
- * User Types
- * --------------------------------------
- * This file defines the official types for user functionality.
+ * Types officiels pour le domaine user.
+ * Toute modification doit être synchronisée dans tous les mocks et composants.
+ * Ne jamais dupliquer ce type en local.
  */
 
 export interface User {
